requirements.txt
# requirements.txt

# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
python-dotenv==1.0.1

# Azure SDK
azure-storage-blob==12.20.0
azure-identity==1.16.0
azure-keyvault-secrets==4.8.0
azure-mgmt-resource==23.0.0

# Data Processing
pandas==2.2.2
numpy==1.26.4
polars==0.20.22
pyarrow==16.0.0
delta-spark==3.1.0

# Machine Learning
scikit-learn==1.4.2
prophet==1.1.5
xgboost==2.0.3
mlflow==2.13.0
shap==0.45.0

# Geospatial
geopandas==0.14.4
folium==0.16.0
shapely==2.0.4
geopy==2.4.1

# HTTP / APIs
httpx==0.27.0
aiohttp==3.9.5
requests==2.32.2

# Dashboard
streamlit==1.35.0
plotly==5.22.0
pydeck==0.9.1

# Qualidade de Dados
great-expectations==0.18.15
pandera==0.19.3

# Testes
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0